server/requirements.txt
# requirements.txt for Python with mypy for development.
#

# packages
aiofiles
mypy
fastapi
fastapi-utils
google-cloud-storage
gunicorn
#../luncho_python  # don't include for GAE
mypy-extensions
pycountry
pycountry-convert
pytest
requests
uvicorn
#orjson does not support Pypy
#ujson does not support Pypy
